global edit

+++ src/main/java/com/example/demo/service/UserService.java
@@ -5,8 +5,8 @@ import com.example.demo.model.User;
 
 public interface UserService {
     void userAdd(String name, String email, String pass);
-    void userEdit(int id, String name, String email, String pass);
-    User getById(int id);
+    void userEdit(Long id, String name, String email, String pass);
+    User getById(Long id);
     void delete(User user);
     Iterable<User> getAll();
 }
